requirement.txt
faiss-cpu   
sentence_transformers
nltk